学生选择指导老师 dto service

+++ j2ee-db/src/main/java/com/j2ee/db/dto/StuTeaChDto.java
@@ -77,7 +77,7 @@ public class StuTeaChDto {
      * 老师是否接受成为导师
      *      0等待回复  1同意  -1不同意
      */
-    Short isAccept;
+    Byte isAccept;
 
     /**
      * 是否逻辑删除
